rec: rename mod action → unit_action
