+++ crates/genie-rec/src/unit.rs
@@ -1,4 +1,4 @@
-use crate::action::UnitAction;
+use crate::unit_action::UnitAction;
 use crate::unit_type::UnitBaseClass;
 use crate::Result;
 use crate::{ObjectID, PlayerID};
